Broke up serializers.rs into its own folder
