+++ tendermint/src/serializers/bytes.rs
@@ -1,0 +1,79 @@
+//! Serialize/deserialize bytes (Vec<u8>) type
+
+/// Serialize into hexstring, deserialize from hexstring
+pub mod hexstring {
+    use serde::{Deserialize, Deserializer, Serializer};
+    use subtle_encoding::hex;
+
+    /// Deserialize hexstring into Vec<u8>
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let string = Option::<String>::deserialize(deserializer)?.unwrap_or_default();
+        hex::decode_upper(&string)
+            .or_else(|_| hex::decode(&string))
+            .map_err(serde::de::Error::custom)
+    }
+
+    /// Serialize from T into hexstring
+    pub fn serialize<S, T>(value: &T, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+        T: AsRef<[u8]>,
+    {
+        let hex_bytes = hex::encode(value.as_ref());
+        let hex_string = String::from_utf8(hex_bytes).map_err(serde::ser::Error::custom)?;
+        serializer.serialize_str(&hex_string)
+    }
+}
+
+/// Serialize into base64string, deserialize from base64string
+pub mod base64string {
+    use serde::{Deserialize, Deserializer, Serializer};
+    use subtle_encoding::base64;
+
+    /// Deserialize base64string into Vec<u8>
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let string = Option::<String>::deserialize(deserializer)?.unwrap_or_default();
+        base64::decode(&string).map_err(serde::de::Error::custom)
+    }
+
+    /// Serialize from T into base64string
+    pub fn serialize<S, T>(value: &T, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+        T: AsRef<[u8]>,
+    {
+        let base64_bytes = base64::encode(value.as_ref());
+        let base64_string = String::from_utf8(base64_bytes).map_err(serde::ser::Error::custom)?;
+        serializer.serialize_str(&base64_string)
+    }
+}
+
+/// Serialize into string, deserialize from string
+pub mod string {
+    use serde::{ser::Error, Deserialize, Deserializer, Serializer};
+
+    /// Deserialize string into Vec<u8>
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let string = Option::<String>::deserialize(deserializer)?.unwrap_or_default();
+        Ok(string.as_bytes().to_vec())
+    }
+
+    /// Serialize from T into string
+    pub fn serialize<S, T>(value: &T, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+        T: AsRef<[u8]>,
+    {
+        let string = String::from_utf8(value.as_ref().to_vec()).map_err(Error::custom)?;
+        serializer.serialize_str(&string)
+    }
+}
